feat: add `Stream` and `Stream.OnStreamComplete` handling
