+++ service/options.go
@@ -1,7 +1,6 @@
 package service
 
 import (
-	extproc "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 	"github.com/getyourguide/extproc-go/filter"
 	"github.com/go-logr/logr"
 	"go.opentelemetry.io/otel/trace"
@@ -27,12 +26,14 @@ func WithLogger(log logr.Logger) Option {
 func WithFilters(filters ...filter.Filter) Option {
 	return optionFunc(func(svc *ExtProcessor) {
 		svc.filters = filters
-	})
-}
 
-func WithOnStreamEndFn(fn func(req *filter.RequestContext, msg *extproc.ProcessingRequest)) Option {
-	return optionFunc(func(svc *ExtProcessor) {
-		svc.onStreamEndFn = fn
+		var streams []filter.Stream
+		for _, f := range filters {
+			if s, ok := f.(filter.Stream); ok {
+				streams = append(streams, s)
+			}
+		}
+		svc.streams = streams
 	})
 }
